pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "app-generator"
version = "0.1.0"
authors = [{ name = "Joshua Oliphant", email = "josholiphantlmp@icloud.com" }]
description = "AI-powered FastAPI application generator"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "anthropic>=0.39.0",
    "black>=24.10.0",
    "datamodel-code-generator>=0.26.3",
    "fastapi>=0.115.5",
    "jinja2>=3.1.4",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
    "uvicorn>=0.32.0",
]

[project.scripts]
generate = "app_generator.cli:generate"
